requirements.txt
streamlit==1.31.1
google-generativeai==0.3.2
Pillow==10.2.0
requests==2.31.0
gTTS==2.5.1
python-dotenv==1.0.1
